+++ packages/analytics/src/analytics.test.ts
@@ -15,7 +15,11 @@ t.describe('Analytics Integration', () => {
   let scope: nock.Scope;
 
   const eventProperties: MMConnectProperties = {
-    mmconnect_version: '1.0.0',
+    mmconnect_version: {
+      'connect-multichain': '0.8.0',
+      'connect-evm': '0.6.0',
+      'connect-solana': '0.4.0',
+    },
     dapp_id: 'aave.com',
     anon_id: 'bbbc1727-8b85-433a-a26a-e9df70ddc81c',
     platform: 'web-desktop',