+++ packages/connect-evm/src/index.ts
@@ -1,3 +1,9 @@
+import { registerPackageVersion } from '@metamask/connect-multichain';
+
+import pkg from '../package.json';
+
+registerPackageVersion('connect-evm', pkg.version);
+
 export { getInfuraRpcUrls } from './utils/infura';
 export { createEVMClient, type MetamaskConnectEVM } from './connect';
 export type { EIP1193Provider } from './provider';