+++ packages/connect-solana/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Register `connect-solana` package version in the analytics version registry at module load ([#206](https://github.com/MetaMask/connect-monorepo/pull/206))
+
 ## [0.4.0]
 
 ### Changed